+++ src/main/java/com/catkatpowered/katserver/tokenpool/KatTokenPool.java
@@ -45,8 +45,8 @@ public class KatTokenPool {
         switch (action) {
             case All -> this.tokenPool.clear();
             case Outdated -> {
-                int outdatedTime = KatServer.KatConfigAPI
-                        .<Integer>getConfig(KatConfigNodeConstants.KAT_CONFIG_TOKENPOOL_OUTDATED)
+                Long outdatedTime = KatServer.KatConfigAPI
+                        .<Long>getConfig(KatConfigNodeConstants.KAT_CONFIG_TOKENPOOL_OUTDATED)
                         .get();
                 for (var item : this.tokenPool.entrySet())
                     if ((System.currentTimeMillis() - item.getValue()) >= outdatedTime)
